@@ -1,8 +1,9 @@
 package be.roots.taconic.pricingguide.domain;
 
-import org.junit.Test;
 
-import static org.junit.Assert.assertEquals;
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 public class ContactTest {
 
