@@ -25,9 +25,10 @@ package be.roots.taconic.pricingguide.util;
  *
  */
 
-import org.junit.Test;
 
-import static org.junit.Assert.assertEquals;
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 public class GreekAlphabetTest {
 
