@@ -1,8 +1,9 @@
 package be.roots.taconic.pricingguide.util;
 
-import org.junit.Test;
 
-import static org.junit.Assert.*;
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.*;
 
 public class ModelUtilTest {
 
